pyproject.toml
[project]
name = "speechlab"
version = "0.1.0"
description = "Lumivox AI speech lab"
authors = [
    {name = "ReanGD", email = "reangd@gmail.com"},
]
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = {text = "Apache-2.0"}
keywords = ["speech", "tts", "stt", "voice", "audio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "ormsgpack>=1.9.0",
    "pydantic==2.10.6",
    "pyzmq==26.3.0",
    "loguru==0.7.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]

[project.optional-dependencies]
cpu = [
    "torch==2.5.1",
    "torchaudio==2.5.1",
]
gpu = [
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "triton==3.1.0",
]
onnx = [
    "onnx==1.17.0",
    "onnxruntime-gpu==1.21.0",
]
tensorrt = [
    "tensorrt==10.9.0.34",
    "pycuda==2025.1",
]
pytorch = [
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
]
server = [
    "numpy==2.1.3",
    "click==8.1.8",
    "tqdm==4.67.1",
    "gradio==5.29.0",
    "librosa==0.11.0",
    "tomlkit==0.13.2",
    "requests==2.32.3",
    "huggingface-hub==0.30.1",
]
fish-speech = [
    "fish-speech==0.1.0",
]
xtts = [
    "coqui-tts==0.1.0",
    "omogre",
    # "deepspeed==0.16.7",
]
ru-norm = [
    "runorm==1.1",
]
docker = [
    "speechlab[pytorch,server,fish-speech,xtts,ru-norm]",
]
develop = [
    "speechlab[gpu,onnx,pytorch,server,fish-speech,xtts,ru-norm]",
]

[project.urls]
Homepage = "https://github.com/LumivoxAI/speechlab"
Documentation = "https://github.com/LumivoxAI/speechlab"
Issues = "https://github.com/LumivoxAI/speechlab/issues"

[tool.uv]
managed = true
conflicts = [
    [
      { extra = "cpu" },
      { extra = "gpu" },
    ],
]

[tool.uv.sources]
torch = [
  { index = "torch-cpu", extra = "cpu" },
  { index = "torch-gpu", extra = "gpu" },
]
fish-speech = { git = "https://github.com/fishaudio/fish-speech", rev = "d9f9f148815cf7f7f5d3690b785f4c1273910097"}
coqui-tts = { git = "https://github.com/idiap/coqui-ai-TTS", rev = "97e4deb58c86dfa222e90af36cf9f2d88dfbcf17"}
omogre = { git = "https://github.com/omogr/omogre.git" }

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name = "tensorrt-llm"
url = "https://pypi.nvidia.com"
explicit = true

[[tool.uv.dependency-metadata]]
name = "fish-speech"
version = "0.1.0" # rev = "d9f9f1", master after 1.5.0
requires-dist = [
    "transformers==4.51.3",
    "natsort>=8.4.0",
    "lightning>=2.1.0",
    "rich>=13.5.3",
    "tiktoken>=0.8.0",
    "einops>=0.7.0",
    "loralib>=0.1.2",
    "vector_quantize_pytorch==1.14.24",
]

[[tool.uv.dependency-metadata]]
name = "coqui-tts"
version = "0.1.0" # rev = "97e4de", dev after v0.26.0
requires-dist = [
    # Core
    "anyascii>=0.3.0",
    # Training
    "matplotlib>=3.8.4",
    # Coqui stack
    "coqui-tts-trainer>=0.2.0,<0.3.0",
    "coqpit-config>=0.2.0,<0.3.0",
    # Gruut
    "gruut>=2.4.0",
    # Tortoise
    "transformers>=4.47.0",
    # XTTS
    "num2words>=0.5.14",
    "spacy>=3.8,<4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
